Put common code used for testing code into datafusion/test_utils
