+++ datafusion/core/Cargo.toml
@@ -105,8 +105,8 @@ csv = "1.1.6"
 ctor = "0.1.22"
 doc-comment = "0.3"
 env_logger = "0.9"
-fuzz-utils = { path = "fuzz-utils" }
 rstest = "0.15.0"
+test-utils = { path = "../../test-utils" }
 
 [[bench]]
 harness = false
