+++ test-utils/Cargo.toml
@@ -16,7 +16,7 @@
 # under the License.
 
 [package]
-name = "fuzz-utils"
+name = "test-utils"
 version = "0.1.0"
 edition = "2021"
